+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2022-10-02
+
+### New
+- Emit `outlet-change` event when an outlet is populated
+
+### Updates
+- Uplift devDependencies
+- Convert e2e tests to Cypress v10.x
+
 ## [3.1.1] - 2021-11-16
 
 ### Fixes